+++ client/src/components/auth/login/index.js
@@ -1,0 +1,114 @@
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import { Form, Button, Row, Col } from "antd";
+import { connect } from "react-redux";
+import { loginUser } from "../../../actions/authActions";
+import TextFieldGroup from "../../common/TextFieldGroup";
+
+import styles from "./Login.module.css";
+
+class Login extends Component {
+  constructor() {
+    super();
+    this.state = {
+      email: "",
+      password: "",
+      errors: {}
+    };
+  }
+
+  componentDidMount() {
+    if (this.props.auth.isAuthenticated) {
+      this.props.history.push("/dashboard");
+    }
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.auth.isAuthenticated) {
+      this.props.history.push("/dashboard");
+    }
+
+    if (nextProps.errors) {
+      this.setState({ errors: nextProps.errors });
+    }
+  }
+
+  onSubmit = e => {
+    debugger;
+    e.preventDefault();
+
+    const userData = {
+      email: this.state.email,
+      password: this.state.password
+    };
+
+    this.props.loginUser(userData);
+  };
+
+  onChange = e => {
+    this.setState({ [e.target.name]: e.target.value });
+  };
+
+  render() {
+    const { errors } = this.state;
+
+    return (
+      <div className="login">
+        <Row type="flex" justify="center">
+          <Col span={24}>
+            <h1 className="display-41 text-center1">Log In</h1>
+            <p className="lead1 text-center1">
+              Sign in to your DevConnector account
+            </p>
+            <Form noValidate onSubmit={this.onSubmit}>
+              <TextFieldGroup
+                size={"large"}
+                placeholder="Email Address"
+                name="email"
+                type="email"
+                value={this.state.email}
+                onChange={this.onChange}
+                error={errors.email}
+              />
+              <TextFieldGroup
+                size={"large"}
+                placeholder="Password"
+                name="password"
+                type="password"
+                value={this.state.password}
+                onChange={this.onChange}
+                error={errors.password}
+              />
+              <Button
+                size="large"
+                className="submit-btn"
+                type="primary"
+                htmlType="submit"
+              >
+                Submit
+              </Button>
+            </Form>
+          </Col>
+        </Row>
+      </div>
+    );
+  }
+}
+
+Login.propTypes = {
+  loginUser: PropTypes.func.isRequired,
+  auth: PropTypes.object.isRequired,
+  errors: PropTypes.object.isRequired
+};
+
+const mapStateToProps = state => ({
+  auth: state.auth,
+  errors: state.errors
+});
+
+const WrappedLogin = Form.create()(Login);
+
+export default connect(
+  mapStateToProps,
+  { loginUser }
+)(WrappedLogin);
